Revert experiments -- ok, I proved to myself...
...that there's no acceptable amount of work you can do *that* many times, and
iterating over the polymer is simply unfeasible. Then I looked up spoilers/hints
on the reddit, and it's a head-slapper now that I think about it: this is the
same as the lanternfish. That is:

- You have to model the sum of the system each step instead of the individual parts of the system.

I was tricked by this one because it's an ordered list instead of a random one!
But actually, the question text only asks you for how many of each element there
are at the end. SIGH.