OK I THINK I FOUND IT: went to reddit for a hint, and it's definitely a problem in the literal values.
https://www.reddit.com/r/adventofcode/comments/rhug8s/day_16_part_2_if_your_test_samples_work_but_your/

> if all your operations seem to  be working correctly, look at your type 4 values

I'm adding discrete 4-bit numbers
instead of extending a run of bits that *then* becomes a number. That's wrong!

A nice basic brain glitch, instead of a real squirrelly problem.

This didn't emerge in the test content, because the examples never use literal
values greater than 15 (the limit for a 4-bit uint). ARGH.