Right, so this comment doesn't quite catch the problem case, but it DOES point the way...
I went looking for a non-spoilery rules clarification, because I was starting to think I
misunderstood something about the reduction rules. And the folks in the reddit seem
to be pretty sure my _original_ interpretation of the rules (before I started working)
was right -- process ALL splodes first, only move on to splits once there are 0 splodes,
and return to splodes whenever necessary.

That's going to blow out the levels of recursion somewhat --
many more steps, at least double!